Handle XDG paths for the database file
The default file name is now "{muc}.db", with "{muc}" substituted with
the name of the current MUC. The substitution done in the user-provided
database file name as well. This makes easy to handle multiple instances
of the bot connecting to several MUCs.

This will also allow one to set up an alternative data directory in the
(future) configuration file, as simply as:
  database = /var/lib/kaabot/{muc}.db

This commit also changes the short option for --database from -db to -b
(like base). This avoids having a single-dash long option.